+++ dds_web/security/auth.py
@@ -18,7 +18,12 @@ import structlog
 
 # Own modules
 from dds_web import basic_auth, auth, mail
-from dds_web.errors import AuthenticationError, AccessDeniedError, InviteError, TokenMissingError
+from dds_web.errors import (
+    AuthenticationError,
+    AccessDeniedError,
+    InviteError,
+    TokenMissingError,
+)
 from dds_web.database import models
 import dds_web.utils
 
@@ -233,6 +238,9 @@ def verify_token(token):
     if not user:
         raise AccessDeniedError(message="Invalid token. Try reauthenticating.")
 
+    # Block all users but Super Admins during maintenance
+    dds_web.utils.block_if_maintenance(user=user)
+
     if user.password_reset:
         token_expired = claims.get("exp")
         token_issued = datetime.datetime.fromtimestamp(token_expired) - MFA_EXPIRES_IN
@@ -396,6 +404,9 @@ def verify_password(username, password):
     user = models.User.query.get(username)
 
     if user and user.is_active and user.verify_password(input_password=password):
+        # Block all users but Super Admins during maintenance
+        dds_web.utils.block_if_maintenance(user=user)
+
         if not user.totp_enabled:
             send_hotp_email(user)
         return user